requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn[standard]>=0.30.0,<1.0.0
supabase>=2.16.0,<2.17.0
postgrest>=1.1.1,<1.2.0
pydantic>=2.10.0,<3.0.0
pydantic-settings>=2.6.0,<3.0.0
python-jose[cryptography]>=3.3.0,<4.0.0
python-multipart>=0.0.20,<1.0.0
httpx>=0.27.0,<1.0.0
email-validator>=2.2.0,<3.0.0
numpy>=1.26,<3
pillow>=11.0,<12.0
scikit-learn>=1.5,<2.0
